+++ src/main/java/raisetech/student/management/data/Student.java
@@ -1,5 +1,4 @@
-package raisetech.student.management;
-
+package raisetech.student.management.data;
 
 import lombok.Getter;
 import lombok.Setter;
