+++ src/main/java/raisetech/student/management/data/StudentCourse.java
@@ -1,4 +1,4 @@
-package raisetech.student.management;
+package raisetech.student.management.data;
 
 import java.util.Date;
 import lombok.Getter;
